refactor: streamline model imports by consolidating from default configuration files
