+++ src/renderer/src/pages/home/Inputbar/MentionModelsButton.tsx
@@ -1,8 +1,7 @@
 import ModelTagsWithLabel from '@renderer/components/ModelTagsWithLabel'
 import { useQuickPanel } from '@renderer/components/QuickPanel'
 import { QuickPanelListItem } from '@renderer/components/QuickPanel/types'
-import { isEmbeddingModel, isRerankModel, isVisionModel } from '@renderer/config/models'
-import { getModelLogo } from '@renderer/config/models'
+import { getModelLogo, isEmbeddingModel, isRerankModel, isVisionModel } from '@renderer/config/models'
 import db from '@renderer/databases'
 import { useProviders } from '@renderer/hooks/useProvider'
 import { getModelUniqId } from '@renderer/services/ModelService'
